Tabla Cuenta Creada en la base de datos

+++ Data/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -224,11 +224,11 @@ namespace PracticaCalificada2.Data.Migrations
 
             modelBuilder.Entity("PracticaCalificada2.Models.CuentaBancaria", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("integer");
+                        .HasColumnType("bigint");
 
-                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("Id"));
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("Id"));
 
                     b.Property<string>("Email")
                         .HasColumnType("text");
