+++ Models/CuentaBancaria.cs
@@ -7,7 +7,7 @@ namespace PracticaCalificada2.Models
     public class CuentaBancaria
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }  // Identificador único de la cuenta
+        public long Id { get; set; }  // Identificador único de la cuenta
         public string? NombreTitular { get; set; }  // Nombre del titular de la cuenta
         public string? TipoCuenta { get; set; }  // Tipo de cuenta (Ej: Ahorros, Corriente)
         public decimal SaldoInicial { get; set; }  // Saldo inicial de la cuenta
